==> 2_procrustes_by_variety/2_ampelometry_plot_each_variety.ipynb <==
# set variety name

variety = "Zinfandel"

# import libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

# record the number of points for each segment for indexing purposes

m_pts = 567
d_pts = 490
p_pts = 361
m1_pts = 341
d1_pts = 299
p1_pts = 223
m2_pts = 256
d2_pts = 233
p2_pts = 187
m3_pts = 177
d3_pts = 156
p3_pts = 138
ma_pts = 247
da_pts = 206
pa_pts = 368
mb_pts = 171
db_pts = 164
pb_pts = 166
mc_pts = 109
dc_pts = 102
pc_pts = 108
md_pts = 177
dd_pts = 144
pd_pts = 137
ds_pts = 333
ps_pts = 139

# check that the sum of the points is correct (5999)

sum_check = m_pts + d_pts + p_pts + m1_pts + d1_pts + p1_pts + m2_pts + d2_pts + p2_pts + m3_pts + d3_pts + p3_pts + ma_pts + da_pts + pa_pts + mb_pts + db_pts + pb_pts + mc_pts + dc_pts + pc_pts + md_pts + dd_pts + pd_pts + ds_pts + ps_pts

print(sum_check)


# read in procrustes adjusted coordinates and the mean leaf

all_leaves = np.loadtxt('Desktop/2_Procrustes_by_variety/output_procrustes_coordinates/' + variety + '_adjusted.txt')
mean_shape = np.loadtxt('Desktop/2_Procrustes_by_variety/output_mean_shapes/' + variety + '_mean.txt')

per_leaf = int(len(all_leaves)/4)

print(per_leaf/2)

# plot out each leaf superimposed on each other

alph = 0.6
col = "gray"
linewidth = 6

for i in range(4):

    leaf = all_leaves[(per_leaf*i):(per_leaf*(i+1))]
    
    leafx = leaf[0:int(per_leaf/2)]
    leafy = leaf[int(per_leaf/2):]

    leafx_m = leafx[0:(m_pts)]
    leafx_p = leafx[m_pts:(m_pts+p_pts)]
    leafx_d = leafx[m_pts+p_pts:(m_pts+p_pts+d_pts)]
    leafx_m1 = leafx[m_pts+p_pts+d_pts:(m_pts+p_pts+d_pts+m1_pts)]
    leafx_p1 = leafx[m_pts+p_pts+d_pts+m1_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts)]
    leafx_d1 = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts)]
    leafx_m2 = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts)]
    leafx_p2 = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts)]
    leafx_d2 = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts)]
    leafx_m3 = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts)]
    leafx_p3 = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts)]
    leafx_d3 = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts)]
    leafx_ma = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts)]
    leafx_pa = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts)]
    leafx_da = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts)]
    leafx_mb = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts)]
    leafx_pb = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts)]
    leafx_db = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts)]
    leafx_mc = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts)]
    leafx_pc = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts)]
    leafx_dc = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts)]
    leafx_md = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts)]
    leafx_pd = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts)]
    leafx_dd = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts)]
    leafx_ps = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts+ps_pts)]
    leafx_ds = leafx[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts+ps_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts+ps_pts+ds_pts)]

    leafy_m = leafy[0:(m_pts)]
    leafy_p = leafy[m_pts:(m_pts+p_pts)]
    leafy_d = leafy[m_pts+p_pts:(m_pts+p_pts+d_pts)]
    leafy_m1 = leafy[m_pts+p_pts+d_pts:(m_pts+p_pts+d_pts+m1_pts)]
    leafy_p1 = leafy[m_pts+p_pts+d_pts+m1_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts)]
    leafy_d1 = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts)]
    leafy_m2 = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts)]
    leafy_p2 = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts)]
    leafy_d2 = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts)]
    leafy_m3 = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts)]
    leafy_p3 = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts)]
    leafy_d3 = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts)]
    leafy_ma = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts)]
    leafy_pa = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts)]
    leafy_da = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts)]
    leafy_mb = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts)]
    leafy_pb = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts)]
    leafy_db = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts)]
    leafy_mc = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts)]
    leafy_pc = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts)]
    leafy_dc = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts)]
    leafy_md = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts)]
    leafy_pd = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts)]
    leafy_dd = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts)]
    leafy_ps = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts+ps_pts)]
    leafy_ds = leafy[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts+ps_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts+ps_pts+ds_pts)]
    
    plt.plot(leafx_m, leafy_m, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_p, leafy_p, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_d, leafy_d, alpha=alph, c=col, lw=linewidth)

    plt.plot(leafx_m1, leafy_m1, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_p1, leafy_p1, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_d1, leafy_d1, alpha=alph, c=col, lw=linewidth)

    plt.plot(leafx_m2, leafy_m2, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_p2, leafy_p2, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_d2, leafy_d2, alpha=alph, c=col, lw=linewidth)

    plt.plot(leafx_m3, leafy_m3, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_p3, leafy_p3, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_d3, leafy_d3, alpha=alph, c=col, lw=linewidth)

    plt.plot(leafx_ma, leafy_ma, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_pa, leafy_pa, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_da, leafy_da, alpha=alph, c=col, lw=linewidth)

    plt.plot(leafx_mb, leafy_mb, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_pb, leafy_pb, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_db, leafy_db, alpha=alph, c=col, lw=linewidth)

    plt.plot(leafx_mc, leafy_mc, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_pc, leafy_pc, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_dc, leafy_dc, alpha=alph, c=col, lw=linewidth)

    plt.plot(leafx_md, leafy_md, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_pd, leafy_pd, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_dd, leafy_dd, alpha=alph, c=col, lw=linewidth)

    plt.plot(leafx_ps, leafy_ps, alpha=alph, c=col, lw=linewidth)
    plt.plot(leafx_ds, leafy_ds, alpha=alph, c=col, lw=linewidth)

    plt.axes().set_aspect('equal', 'box')
    
fig = plt.gcf().set_size_inches((20, 20))
plt.axis('off')
plt.show()
plt.draw()


# plot out mean leaf shape

mean_m = mean_shape[0:(m_pts-1),:]
mean_p = mean_shape[m_pts:(m_pts+p_pts-1),:]
mean_d = mean_shape[m_pts+p_pts:(m_pts+p_pts+d_pts-1),:]
mean_m1 = mean_shape[m_pts+p_pts+d_pts:(m_pts+p_pts+d_pts+m1_pts-1),:]
mean_p1 = mean_shape[m_pts+p_pts+d_pts+m1_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts-1),:]
mean_d1 = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts-1),:]
mean_m2 = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts-1),:]
mean_p2 = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts-1),:]
mean_d2 = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts-1),:]
mean_m3 = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts-1),:]
mean_p3 = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts-1),:]
mean_d3 = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts-1),:]
mean_ma = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts-1),:]
mean_pa = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts-1),:]
mean_da = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts-1),:]
mean_mb = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts-1),:]
mean_pb = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts-1),:]
mean_db = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts-1),:]
mean_mc = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts-1),:]
mean_pc = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts-1),:]
mean_dc = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts-1),:]
mean_md = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts-1),:]
mean_pd = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts-1),:]
mean_dd = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts-1),:]
mean_ps = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts+ps_pts-1),:]
mean_ds = mean_shape[m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts+ps_pts:(m_pts+p_pts+d_pts+m1_pts+p1_pts+d1_pts+m2_pts+p2_pts+d2_pts+m3_pts+p3_pts+d3_pts+ma_pts+pa_pts+da_pts+mb_pts+pb_pts+db_pts+mc_pts+pc_pts+dc_pts+md_pts+pd_pts+dd_pts+ps_pts+ds_pts-1),:]

pink = "#B31A82"
orange = "#F28317"
size = 800 # point size
linewidth = 6 # line width
fs = 40 # font size
ls = 40 # Tick number font size
a = 1

plt.plot(mean_m[:,0], mean_m[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_p[:,0], mean_p[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_d[:,0], mean_d[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_m1[:,0], mean_m1[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_p1[:,0], mean_p1[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_d1[:,0], mean_d1[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_m2[:,0], mean_m2[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_p2[:,0], mean_p2[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_d2[:,0], mean_d2[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_m3[:,0], mean_m3[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_p3[:,0], mean_p3[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_d3[:,0], mean_d3[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_ma[:,0], mean_ma[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_pa[:,0], mean_pa[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_da[:,0], mean_da[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_mb[:,0], mean_mb[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_pb[:,0], mean_pb[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_db[:,0], mean_db[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_mc[:,0], mean_mc[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_pc[:,0], mean_pc[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_dc[:,0], mean_dc[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_md[:,0], mean_md[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_pd[:,0], mean_pd[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_dd[:,0], mean_dd[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_ps[:,0], mean_ps[:,1], lw=linewidth, c=pink, alpha=a)
plt.plot(mean_ds[:,0], mean_ds[:,1], lw=linewidth, c=pink, alpha=a)

### Branch points

plt.scatter(mean_m[0,0], mean_m[0,1], s=size, c=orange, alpha=a)

plt.scatter(mean_m1[0,0], mean_m1[0,1], s=size, c=orange, alpha=a)
plt.scatter(mean_d1[0,0], mean_d1[0,1], s=size, c=orange, alpha=a)
plt.scatter(mean_p1[0,0], mean_p1[0,1], s=size, c=orange, alpha=a)

plt.scatter(mean_m2[0,0], mean_m2[0,1], s=size, c=orange, alpha=a)
plt.scatter(mean_d2[0,0], mean_d2[0,1], s=size, c=orange, alpha=a)
plt.scatter(mean_p2[0,0], mean_p2[0,1], s=size, c=orange, alpha=a)

plt.scatter(mean_m3[0,0], mean_m3[0,1], s=size, c=orange, alpha=a)
plt.scatter(mean_d3[0,0], mean_d3[0,1], s=size, c=orange, alpha=a)
plt.scatter(mean_p3[0,0], mean_p3[0,1], s=size, c=orange, alpha=a)

### Blade points

plt.scatter(mean_ma[-1,0], mean_ma[-1,1], s=size, c=orange, alpha=a)
plt.scatter(mean_da[-1,0], mean_da[-1,1], s=size, c=orange, alpha=a)
plt.scatter(mean_pa[-1,0], mean_pa[-1,1], s=size, c=orange, alpha=a)

plt.scatter(mean_mb[-1,0], mean_mb[-1,1], s=size, c=orange, alpha=a)
plt.scatter(mean_db[-1,0], mean_db[-1,1], s=size, c=orange, alpha=a)
plt.scatter(mean_pb[-1,0], mean_pb[-1,1], s=size, c=orange, alpha=a)

plt.scatter(mean_mc[-1,0], mean_mc[-1,1], s=size, c=orange, alpha=a)
plt.scatter(mean_dc[-1,0], mean_dc[-1,1], s=size, c=orange, alpha=a)
plt.scatter(mean_pc[-1,0], mean_pc[-1,1], s=size, c=orange, alpha=a)

plt.scatter(mean_md[-1,0], mean_md[-1,1], s=size, c=orange, alpha=a)
plt.scatter(mean_dd[-1,0], mean_dd[-1,1], s=size, c=orange, alpha=a)
plt.scatter(mean_pd[-1,0], mean_pd[-1,1], s=size, c=orange, alpha=a)

plt.scatter(mean_ds[-1,0], mean_ds[-1,1], s=size, c=orange, alpha=a)
plt.scatter(mean_ps[-1,0], mean_ps[-1,1], s=size, c=orange, alpha=a)

plt.gcf().set_size_inches((20, 20))
plt.axes().set_aspect('equal', 'box')
plt.axis('off')
plt.show()
